feat(wayland/connection): Make `connection`-class virtual so we can extend it's functionality
This introduces three new virtual methods: `on_resize()`, `on_frame()` and `on_close()`
as well as two getters for the `wl_display` and `wl_surface` whis connection created.

These callbacks will be called in the appropriate wayland-callbacks; The getters can
be used by subclasses to access both handles in case they need them for further initialization
or drawing.

This way we can keep the responsibilities of the `connection`-class at initializing
the wayland connection and surface as well as handling the communication while implementation
or platform (think EGL, Vulkan) specific details can be implemented in a subclass.

One issue with the virtual dispatch is that the `connection`'s constructor used to do
a wayland-roundtrip after first initialization of the surface. This doesn't work
anymore because the wayland-callbacks that would be triggered by that roundtrip depend
on unitialized virtual methods (when instantiating a subclass the constructor of the
virtual base class is called first - meaning that the vtable of the subclass was not
yet set up).

Maybe just leave the first roundtrip to the user...